Move scanner lib files to separate include and src dir
Create cmakelists that creates/adds external libraries to the project (pfd and crc)

+++ lib/src/filepath.cpp
@@ -34,7 +34,7 @@ void FilePath::setRoot(const std::string& root_, uint16_t crc)
     partCrcs.clear();
     parts.push_back(root);
     if(crc == 0)
-        crc = crc16((char*)root.c_str(), root.length());
+        crc = crc16((char*)root.c_str(), (uint16_t) root.length());
     partCrcs.push_back(crc);
 }
 
@@ -109,7 +109,7 @@ bool FilePath::addDir(const std::string& name, uint16_t crc)
         parts.back().push_back(PlatformUtils::filePathSeparator);
 
     if(crc == 0)
-        crc = crc16((char*)name.c_str(), name.length());
+        crc = crc16((char*)name.c_str(),(uint16_t) name.length());
     partCrcs.push_back(crc);
     return true;
 }
@@ -128,7 +128,7 @@ bool FilePath::addFile(const std::string& name, uint16_t crc)
     }
     parts.push_back(name);
     if(crc == 0)
-        crc = crc16((char*)name.c_str(), name.length());
+        crc = crc16((char*)name.c_str(), (uint16_t) name.length());
     partCrcs.push_back(crc);
     return true;
 }
